requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
motor==3.3.2
pymongo==4.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.4.2
pydantic-settings==2.0.3
python-dateutil==2.8.2
httpx==0.25.2
pytest==7.4.3
pytest-asyncio==0.21.1
